scripts/requirements.txt
pyspark
google
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-dotenv
pandas
influxdb
influxdb-client